presentation-decks/2024/Track2-AaronSaikovski-Python for the Azure Engineer/demo/pyproject.toml
[tool.ruff] 
extend-select = [
    "B",
    "BLE",
    "C4",
    "ERA",
    "I",
    "ICN",
    "INP",
    "ISC",
    "N",
    "NPY",
    "PGH",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TCH",
    "TID",
    "YTT",
]
ignore=["E402","B905","N818","S101","T201"]
show-fixes = true
[tool.poetry]
name = "azbootcamp-2024-python-demos"
version = "0.1.0"
description = "Global Azure Bootcamp 2024 Python Demos"
authors = ["asaikovski@outlook.com"]
license = "MIT"
readme = "README.md"
packages = [{include = "azbootcamp_2024_python_demos"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
azure-mgmt-resource = "^23.0.1"
azure-identity = "^1.16.0"
azure-mgmt-storage = "^21.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
black = "^24.4.0"
ruff = "^0.3.7"
bandit = "^1.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
